% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Flagstaff Weather at Pulliam Airport}
\format{
A data frame with 29,449 observations and 5 columns.
\describe{
   \item{DATE}{The date of each entry in the lubridate ymd format}
   \item{PRCP}{Precipitation in inches}
   \item{SNOW}{Snow in inches}
   \item{TMAX}{Max temperature of that day in Fahrenheit}
   \item{TMIN}{Min temperature of that day in Fahrenheit}
}
}
\usage{
Flagstaff_Weather
}
\description{
This data contains information about the weather at Flagstaff AZ at the
Pulliam Airport.The data includes the Dates from 1950 to 2019. It also
contains the information about precipitation, snow, and temperature maximum
and minimum.
}
\keyword{datasets}
